4.21 refactor DELETE, only user who added blog can delete, token check

+++ controllers/blogs.js
@@ -34,8 +34,17 @@ blogsRouter.post('/', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  const requesterId = decodedToken.id;
+
+  const blog = await Blog.findById(request.params.id);
+  if (blog.user.toString() === requesterId) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  } else {
+    response.status(401).json({ error: 'invalid user' });
+  }
 });
 
 blogsRouter.put('/:id', async (request, response) => {
